+++ rust/pyo3-geoarrow/src/array.rs
@@ -63,6 +63,22 @@ impl PyGeoArray {
                 self.__arrow_c_array__(py, None)?,
             )
     }
+
+    /// Export to a geoarrow.rust.core.GeoArrowArray.
+    ///
+    /// This requires that you depend on geoarrow-rust-core from your Python package.
+    pub fn into_geoarrow_py(self, py: Python) -> PyResult<Bound<PyAny>> {
+        let geoarrow_mod = py.import(intern!(py, "geoarrow.rust.core"))?;
+        let array_capsules = to_array_pycapsules(
+            py,
+            self.0.data_type().to_field("", true).into(),
+            &self.0.to_array_ref(),
+            None,
+        )?;
+        geoarrow_mod
+            .getattr(intern!(py, "GeoArrowArray"))?
+            .call_method1(intern!(py, "from_arrow_pycapsule"), array_capsules)
+    }
 }
 
 #[pymethods]